PM / AVS: SmartReflex: init driver using DTB
Patch adds posibility to init SmartReflex driver using Device Tree.
It updates init sequence in the following way:
- SmartReflex platform data is retrieved from DTB
- NVALUES are read from EFUSES during driver probe
- All calibration algorithm specific data is retrieved from DTB

To keep backward compatibility, SmartReflex can be initialized
in the usual way - using platform_device and platform_data, which are
defined in sr_device.c. This path is used if DTB is not defined for
corresponding SmartReflex.

Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>